Added some validation/verification on the delete function in ImageController. Deleted a duplicate in the main page. Prepared an error page to redirect the user there if twisting with the id input.

+++ src/Controller/ImageController.php
@@ -123,12 +123,18 @@ final class ImageController extends AbstractController
      */
     public function delete(): void
     {
-        // ! TODO verify that ID exists
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+        if (
+            $_SERVER['REQUEST_METHOD'] === 'POST'
+            && isset($_POST['id'])
+            && ctype_digit($_POST['id'])
+        ) {
             $id = trim($_POST['id']);
             $imageManager = new ImageManager();
             $imageManager->delete((int)$id);
 
+            header('Location:/images');
+        } else {
+            // TODO : redirect to /error page
             header('Location:/images');
         }
     }
